requirements.txt
numpy==1.18.0
pandas==1.0.1
Pillow==7.0.0
plotly==4.7.1
streamlit==0.58.0
tensorflow==1.15.0
Keras==2.2.5